+++ pages/_app.tsx
@@ -1,11 +1,21 @@
 import { useMemo, useCallback, useEffect } from 'react'
 import { AppProps } from 'next/app'
+import NextLink from 'next/link'
 import { DefaultSeo, DefaultSeoProps } from 'next-seo'
 import { useRouter } from 'next/router'
-import { Locale, defaultLocale, extractLocaleFromPath, I18nProvider, ThemeProvider } from '@edgeandnode/components'
+import {
+  Layout,
+  NavigationMarketing,
+  Footer,
+  Locale,
+  defaultLocale,
+  extractLocaleFromPath,
+  I18nProvider,
+  ThemeProvider,
+  LocaleSwitcher,
+} from '@edgeandnode/components'
 import '@edgeandnode/components/build/components.css'
 
-import { Layout } from '@/layout'
 import { supportedLocales, translations } from '@/i18n'
 
 const seo: DefaultSeoProps = {
@@ -54,6 +64,8 @@ const MyApp = ({ Component, pageProps }: AppProps) => {
 
   seo.openGraph!.locale = locale
 
+  const localeSwitcher = useMemo(() => <LocaleSwitcher key="localeSwitcher" />, [])
+
   return (
     <I18nProvider
       supportedLocales={supportedLocales}
@@ -64,8 +76,31 @@ const MyApp = ({ Component, pageProps }: AppProps) => {
     >
       <ThemeProvider>
         <DefaultSeo {...seo} />
-        <Layout>
-          <Component {...pageProps} />
+        <div
+          sx={{
+            position: 'absolute',
+            left: 0,
+            right: 0,
+            top: 0,
+            minHeight: '768px',
+            backgroundImage: `url('${process.env.BASE_PATH}/img/page-background.png')`,
+            backgroundSize: 'cover',
+            backgroundPosition: 'center top',
+            '@media (min-width: 1440px)': {
+              aspectRatio: '1440/768',
+            },
+          }}
+        />
+        <Layout
+          mainContainer
+          headerContent={
+            <NavigationMarketing activeRoute="/docs" NextLink={NextLink} rightAlignItems={[localeSwitcher]} />
+          }
+          footerContent={<Footer localeSwitcher={localeSwitcher} />}
+        >
+          <div>
+            <Component {...pageProps} />
+          </div>
         </Layout>
       </ThemeProvider>
     </I18nProvider>
